api update

+++ src/donations/donations.service.ts
@@ -1,35 +1,47 @@
-import { Injectable } from '@nestjs/common';
-import { Prisma } from '@prisma/client';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
+import { Prisma, UserRolesEnum } from '@prisma/client';
 import { PrismaService } from 'src/prisma/prisma.service';
-import { DonationType } from './donation.type';
+import { UserPayloadType } from 'src/auth/types/jwt-payload.type';
+import { UsersService } from 'src/users/users.service';
 
 @Injectable()
 export class DonationsService {
-  private readonly include = { user: true };
+  private readonly include: Prisma.DonationInclude = { user: true };
 
-  constructor(private readonly prisma: PrismaService) {}
+  constructor(
+    private readonly prisma: PrismaService,
+    private readonly usersService: UsersService,
+  ) {}
 
-  async create(
-    data: Prisma.DonationUncheckedCreateInput,
-  ): Promise<DonationType> {
+  async create(data: Prisma.DonationUncheckedCreateInput) {
     return this.prisma.donation.create({ data, include: this.include });
   }
 
-  async findOne(where: Prisma.DonationWhereUniqueInput): Promise<DonationType> {
-    return this.prisma.donation.findUniqueOrThrow({
+  async findOne(
+    userPayload: UserPayloadType,
+    where: Prisma.DonationWhereInput = {},
+  ) {
+    await this._filterDonationsAccess(userPayload, where);
+    return this.prisma.donation.findFirstOrThrow({
       where,
       include: this.include,
     });
   }
 
-  async findMany(where?: Prisma.DonationWhereInput): Promise<DonationType[]> {
+  async findMany(
+    userPayload: UserPayloadType,
+    where: Prisma.DonationWhereInput = {},
+  ) {
+    await this._filterDonationsAccess(userPayload, where);
     return this.prisma.donation.findMany({ where, include: this.include });
   }
 
   async update(
-    where: Prisma.DonationWhereUniqueInput,
+    userPayload: UserPayloadType,
     data: Prisma.DonationUpdateInput,
-  ): Promise<DonationType> {
+    where: Prisma.DonationWhereUniqueInput = {},
+  ) {
+    await this._filterDonationsAccess(userPayload, where);
     return this.prisma.donation.update({
       data,
       where,
@@ -37,7 +49,26 @@ export class DonationsService {
     });
   }
 
-  async delete(where: Prisma.DonationWhereUniqueInput): Promise<DonationType> {
+  async delete(
+    userPayload: UserPayloadType,
+    where: Prisma.DonationWhereUniqueInput = {},
+  ) {
+    if (userPayload.role !== UserRolesEnum.SYSTEM_ADMIN) {
+      throw new UnauthorizedException();
+    }
     return this.prisma.donation.delete({ where, include: this.include });
+  }
+
+  private async _filterDonationsAccess(
+    userPayload: UserPayloadType,
+    where: Prisma.DonationWhereInput = {},
+  ) {
+    if (userPayload.role === UserRolesEnum.DONOR) {
+      where.userId = userPayload.id;
+    }
+    if (userPayload.role === UserRolesEnum.CENTER_ADMIN) {
+      const user = await this.usersService.findOne({ id: userPayload.id });
+      where.centerId = user.employedCenterId;
+    }
   }
 }
